--- README.md
@@ -1,14 +1,0 @@
-# 미션 - 자동차 경주 게임
-
-## 🚀 기능 요구사항
-- [X] 경주할 자동차를 입력하여 생성하는 기능
-- [X] 자동차 이름 입력 시 발생할 수 있는 예외상항 처리
-    - 이름이 6자 이상 또는 0자 이하이면 허용하지 않고 중복 이름일 경우도 허용하지 않는다
-- [X] 자동차 경주 횟수 입력하는 기능
-- [X] 자동차 경주 횟수 입력 시 발생할 수 있는 예외상항 처리
-    - 숫자가 아닌 경우는 허용하지 않는다
-- [X] 시도 횟수별 자동차 이동 거리 누적하는 기능
-    - 전진하는 조건은 0에서 9 사이에서 random 값을 구한 후 random 값이 4 이상일 경우 전진하고, 3 이하의 값이면 멈춘다
-- [X] 시도 횟수별 자동차 누적 이동 거리 출력하는 기능
-- [X] 최종 우승자 지정해 출력하는 기능
-    - 우승자는 한 명 이상일 수 있다
